fast_as/pyproject.toml
[project]
name = "fast-as"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.15.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi[standard]>=0.115.12",
    "passlib>=1.7.4",
    "psycopg2>=2.9.10",
    "pydantic>=2.11.4",
    "pydantic-settings>=2.9.1",
    "pyjwt>=2.10.1",
    "redis>=6.1.0",
    "sqlmodel>=0.0.24",
]
